Implemented better logging and progress bars.
Changed logging and progress bars to be implemented by a protocol and
moved logging to loguru. This should make the system significantly more
flexible in future and allow us to change loggers/progress bars
on-the-fly (which will be useful for developing dashboards built around
this library). For now, we follow the std.logging protocol for logging
and a stripped-down version of tqdm's protocol for progress bars
(because i really like how tqdm handles this). This lays groundwork for
future developemnt using MTPy as a core for our other projects.