Fixed incorrect import in mtpy.base.feedback
Fixed an incorrect import which is (as far as i can tell) the last
addressable cause of significant type instability. The remaining main
causes of type instability now seem to be holoviz and fsspec. Since
fsspec is currently in the process of gradually adding typing this
should improve significantly over time.