Major cleanup of filesystem
Previous iterations left a lot of useless junk and strant empty branches
in the filestructure. This has now been fixed and should simplify
future development for this module. Now that it's modular and cleaner we
should be able to easily develop separate plugin modules to expand the
usefulness of this framework going forward.